Sleep and continue when same target and base index

+++ Mimir.Worker/StateDocumentConverter/AddressStatePair.cs
@@ -5,7 +5,7 @@ namespace Mimir.Worker.StateDocumentConverter;
 
 public class AddressStatePair
 {
-    public int BlockIndex { get; set; }
+    public long BlockIndex { get; set; }
     public Address Address { get; set; }
     public IValue RawState { get; set; }
 }
